refactor(dashrate): extract App into components, hooks, and lib modules
Break the monolithic App.tsx into focused view components, data hooks,
and shared session/ratings modules to reduce its size and improve
maintainability.
