+++ example-apps/dashrate/src/session/types.ts
@@ -1,0 +1,7 @@
+import type { DashKeyManager, DashSdk } from "../dash/types";
+
+export interface Session {
+  sdk: DashSdk;
+  keyManager: DashKeyManager;
+  identityId: string;
+}
